+++ src/components/pages/SignUp/index.tsx
@@ -3,7 +3,7 @@ import { AuthComponent } from '../../organisms/Auth/index';
 import Paper from '@material-ui/core/Paper';
 import { PaperContainer, Title } from './style';
 
-export function SignUpRouter() {
+export function SignUp() {
   return (
     <>
       <Title>Book Manager</Title>
